revert one of the previous commits, instead make code jdk8 compliant
