+++ src/main/java/tech/units/indriya/AbstractQuantity.java
@@ -33,6 +33,8 @@ import static javax.measure.Quantity.Scale.ABSOLUTE;
 
 import java.math.BigDecimal;
 import java.util.Objects;
+import java.util.Optional;
+import java.util.function.UnaryOperator;
 
 import javax.measure.Quantity;
 import javax.measure.Unit;
@@ -42,7 +44,9 @@ import javax.measure.quantity.Dimensionless;
 
 import tech.units.indriya.format.SimpleQuantityFormat;
 import tech.units.indriya.format.SimpleUnitFormat;
+import tech.units.indriya.function.AbstractConverter;
 import tech.units.indriya.function.Calculus;
+import tech.units.indriya.internal.function.Calculator;
 import tech.units.indriya.quantity.Quantities;
 import tech.units.indriya.spi.NumberSystem;
 import tech.uom.lib.common.function.UnitSupplier;
@@ -194,6 +198,13 @@ public abstract class AbstractQuantity<Q extends Quantity<Q>> implements Compara
     public ComparableQuantity<Q> to(Unit<Q> anotherUnit) {
         if (anotherUnit.equals(this.getUnit())) {
             return this;
+        }
+        if (Scale.RELATIVE.equals(this.getScale())) {
+            final Unit<Q> systemUnit = this.getUnit().getSystemUnit();
+            final ToSystemUnitConverter converter = ToSystemUnitConverter.forQuantity(this, systemUnit);
+            final Number valueInSystemUnit = converter.apply(getValue());
+            final Number valueInOtherUnit = systemUnit.getConverterTo(anotherUnit).convert(valueInSystemUnit);
+            return Quantities.getQuantity(valueInOtherUnit, anotherUnit);
         }
         UnitConverter t = getUnit().getConverterTo(anotherUnit);
         Number convertedValue = t.convert(getValue());
@@ -368,5 +379,79 @@ public abstract class AbstractQuantity<Q extends Quantity<Q>> implements Compara
 
     protected NumberSystem numberSystem() {
         return Calculus.currentNumberSystem();
+    }
+    
+    // also honors RELATIVE scale
+    protected static class ToSystemUnitConverter implements UnaryOperator<Number> {
+        private final UnaryOperator<Number> unaryOperator;
+        private final UnaryOperator<Number> inverseOperator;
+        
+        public static <Q extends Quantity<Q>>  
+        ToSystemUnitConverter forQuantity(Quantity<Q> quantity, Unit<Q> systemUnit) {
+            if(quantity.getUnit().equals(systemUnit)) {
+                return ToSystemUnitConverter.noop(); // no conversion required
+            }
+            
+            final UnitConverter converter = quantity.getUnit().getConverterTo(systemUnit);
+            
+            if(ABSOLUTE.equals(quantity.getScale())) {
+                
+                return ToSystemUnitConverter.of(converter::convert); // convert to system units
+                
+            } else {
+                final Number linearFactor = linearFactorOf(converter).orElse(null);
+                if(linearFactor!=null) {
+                    // conversion by factor required ... Δ2°C -> Δ2K , Δ2°F -> 5/9 * Δ2K
+                    return ToSystemUnitConverter.factor(linearFactor); 
+                }
+                // convert any other cases of RELATIVE scale to system unit (ABSOLUTE) ...
+                throw unsupportedConverter(converter, quantity.getUnit());
+            }
+        }
+        
+        public Number invert(Number x) {
+            return isNoop() 
+                    ? x
+                    : inverseOperator.apply(x); 
+        }
+
+        public static ToSystemUnitConverter of(UnaryOperator<Number> unaryOperator) {
+            return new ToSystemUnitConverter(unaryOperator, null);
+        }
+        public static ToSystemUnitConverter noop() {
+            return new ToSystemUnitConverter(null, null);
+        }
+        public static ToSystemUnitConverter factor(Number factor) {
+            return new ToSystemUnitConverter(
+                    number->Calculator.of(number).multiply(factor).peek(),
+                    number->Calculator.of(number).divide(factor).peek());
+        }
+        private ToSystemUnitConverter(UnaryOperator<Number> unaryOperator, UnaryOperator<Number> inverseOperator) {
+            this.unaryOperator = unaryOperator;
+            this.inverseOperator = inverseOperator;
+        }
+        public boolean isNoop() {
+            return unaryOperator==null;
+        }
+        @Override
+        public Number apply(Number x) {
+            return isNoop() 
+                    ? x
+                    : unaryOperator.apply(x); 
+        }
+        
+        private static Optional<Number> linearFactorOf(UnitConverter converter) {
+            return (converter instanceof AbstractConverter)
+                    ? ((AbstractConverter)converter).linearFactor()
+                    : Optional.empty();
+        }
+        
+        private static UnsupportedOperationException unsupportedConverter(UnitConverter converter, Unit<?> unit) {
+            return new UnsupportedOperationException(
+                    String.format(
+                            "Scale conversion from RELATIVE to ABSOLUTE for Unit %s having Converter %s is not implemented.", 
+                            unit, converter));
+        }
+        
     }
 }
